requirements.txt
langchain==0.1.8
langchain_community==0.0.21
PyYAML==6.0
unstructured==0.12.4
